[Implement createMeet]

+++ back-end/quick-meet/src/main/java/ru/imo/quickmeet/dto/ReturnedMeetDTO.java
@@ -3,5 +3,5 @@ package ru.imo.quickmeet.dto;
 
 import java.util.List;
 
-public record ReturnedMeetDTO(int meet_id, List<String> users, long time_start, long duration, String comment) {
+public record ReturnedMeetDTO(long meet_id, List<String> users, long time_start, long duration, String comment) {
 }
